Changes to word censors in the admin panel - Deletions now require confirmation - A few miscellaneous bug fixes Note to translators: This adds a new language variable
git-svn-id: file:///svn/phpbb/branches/phpBB-2_0_0@5493 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/admin/admin_words.php
@@ -20,14 +20,14 @@
  *
  ***************************************************************************/
 
-define('IN_PHPBB', 1);
-
 if( !empty($setmodules) )
 {
 	$file = basename(__FILE__);
 	$module['General']['Word_Censor'] = $file;
 	return;
 }
+
+define('IN_PHPBB', 1);
 
 //
 // Load default header
@@ -38,7 +38,7 @@ require('./pagestart.' . $phpEx);
 
 if( isset($HTTP_GET_VARS['mode']) || isset($HTTP_POST_VARS['mode']) )
 {
-	$mode = ($HTTP_GET_VARS['mode']) ? $HTTP_GET_VARS['mode'] : $HTTP_POST_VARS['mode'];
+	$mode = (isset($HTTP_GET_VARS['mode'])) ? $HTTP_GET_VARS['mode'] : $HTTP_POST_VARS['mode'];
 	$mode = htmlspecialchars($mode);
 }
 else 
@@ -60,6 +60,9 @@ else
 	}
 }
 
+// Restrict mode input to valid options
+$mode = ( in_array($mode, array('add', 'edit', 'save', 'delete')) ) ? $mode : '';
+
 if( $mode != "" )
 {
 	if( $mode == "edit" || $mode == "add" )
@@ -70,6 +73,7 @@ if( $mode != "" )
 			"body" => "admin/words_edit_body.tpl")
 		);
 
+		$word_info = array('word' => '', 'replacement' => '');
 		$s_hidden_fields = '';
 
 		if( $mode == "edit" )
@@ -158,7 +162,9 @@ if( $mode != "" )
 			$word_id = 0;
 		}
 
-		if( $word_id )
+		$confirm = isset($HTTP_POST_VARS['confirm']);
+
+		if( $word_id && $confirm )
 		{
 			$sql = "DELETE FROM " . WORDS_TABLE . " 
 				WHERE word_id = $word_id";
@@ -171,6 +177,26 @@ if( $mode != "" )
 			$message = $lang['Word_removed'] . "<br /><br />" . sprintf($lang['Click_return_wordadmin'], "<a href=\"" . append_sid("admin_words.$phpEx") . "\">", "</a>") . "<br /><br />" . sprintf($lang['Click_return_admin_index'], "<a href=\"" . append_sid("index.$phpEx?pane=right") . "\">", "</a>");
 
 			message_die(GENERAL_MESSAGE, $message);
+		}
+		elseif( $word_id && !$confirm)
+		{
+			// Present the confirmation screen to the user
+			$template->set_filenames(array(
+				'body' => 'admin/confirm_body.tpl')
+			);
+
+			$hidden_fields = '<input type="hidden" name="mode" value="delete" /><input type="hidden" name="id" value="' . $word_id . '" />';
+
+			$template->assign_vars(array(
+				'MESSAGE_TITLE' => $lang['Confirm'],
+				'MESSAGE_TEXT' => $lang['Confirm_delete_word'],
+
+				'L_YES' => $lang['Yes'],
+				'L_NO' => $lang['No'],
+
+				'S_CONFIRM_ACTION' => append_sid("admin_words.$phpEx"),
+				'S_HIDDEN_FIELDS' => $hidden_fields)
+			);
 		}
 		else
 		{
@@ -193,6 +219,7 @@ else
 	}
 
 	$word_rows = $db->sql_fetchrowset($result);
+	$db->sql_freeresult($result);
 	$word_count = count($word_rows);
 
 	$template->assign_vars(array(
